requirements/prod.txt
# requirements/prod.txt
-r prod.txt
Django==1.9.5
pymysql==0.7.3
django-ckeditor==5.0.3
Pillow==3.1.2
